Refactor codes into modules
